Begin reworking import page for optional inference
