+++ mathesar_ui/src/pages/import-upload/column-type-inference/LabelInferenceFalse.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import NameWithIcon from '@mathesar/components/NameWithIcon.svelte';
+  import { iconUiTypeText } from '@mathesar/icons';
+</script>
+
+<NameWithIcon icon={iconUiTypeText} truncate={false}>
+  Use "Text" for all column types
+  <em>(faster)</em>
+</NameWithIcon>
