+++ mathesar_ui/src/pages/import-upload/column-type-inference/LabelInferenceTrue.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import NameWithIcon from '@mathesar/components/NameWithIcon.svelte';
+  import { iconInferColumnType } from '@mathesar/icons';
+</script>
+
+<NameWithIcon icon={iconInferColumnType} truncate={false}>
+  Analyze data and pick the best type for each column
+  <em>(slower)</em>
+</NameWithIcon>
